Adds cache to save pulled answers
Added temp cache directory for file cache storage (maybe sqlite database storage
at some point?)
Renamed and unified files that found answers so only one is used and the
parameter is evaluated to determine the day of the answer.
Tried to streamline the libraries to isolate them and create clear boundaries.

Trying to work on software architecture/design at the same time as working on
racket as the language.