Adds answer search; tries to improve structure
Added a tool to search all answers and tell the user when/if the word
appeared and when (if it did/will).
Tried to streamline the wordle answer pull because the url is known and
several aspects are shared (more to go? looking at you, command-line!).
Updated documentation.

This is an interesting tool but is also a stepping stone to creating
guessers that are interchangable (i.e. with different guessing
strategies). It will be helpful to know if an answer has appeared (or
will appear, perhaps?).